+++ src/main/java/com/cburch/logisim/soc/gui/AssemblerPanel.java
@@ -16,6 +16,7 @@ import com.cburch.contracts.BaseKeyListenerContract;
 import com.cburch.contracts.BaseMouseListenerContract;
 import com.cburch.contracts.BaseWindowListenerContract;
 import com.cburch.logisim.circuit.CircuitState;
+import com.cburch.logisim.gui.generic.EditorTheme;
 import com.cburch.logisim.gui.generic.OptionPane;
 import com.cburch.logisim.gui.icons.CompileIcon;
 import com.cburch.logisim.gui.icons.ErrorIcon;
@@ -28,7 +29,6 @@ import com.cburch.logisim.soc.util.Assembler;
 import com.cburch.logisim.soc.util.AssemblerInterface;
 import com.cburch.logisim.util.LocaleListener;
 import java.awt.BorderLayout;
-import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -152,7 +152,6 @@ public class AssemblerPanel extends JPanel
     info.add(Box.createHorizontalStrut(5));
     info.add(helpLabel);
     info.add(Box.createHorizontalGlue());
-    lineLabel.setOpaque(true);
     info.add(lineLabel);
     info.setPreferredSize(new Dimension(40, AppPreferences.getScaled(20)));
     setLayout(new BorderLayout());
@@ -162,6 +161,7 @@ public class AssemblerPanel extends JPanel
     this.assembler = new Assembler(assembler, debugScrollPane);
     asmWindow.addParser(this.assembler);
     localeChanged();
+    EditorTheme.install(asmWindow);
   }
 
   private void openFile() {
@@ -262,7 +262,6 @@ public class AssemblerPanel extends JPanel
   }
 
   private void updateLineNumber() {
-    lineLabel.setBackground(documentChanged ? Color.YELLOW : Color.WHITE);
     lineLabel.setText(S.get("RV32imAsmLineIndicator", lineNumber, numberOfLines));
     lineLabel.repaint();
   }