+++ src/main/java/com/cburch/logisim/vhdl/gui/HdlContentEditor.java
@@ -12,6 +12,7 @@ package com.cburch.logisim.vhdl.gui;
 import static com.cburch.logisim.vhdl.Strings.S;
 
 import com.cburch.contracts.BaseDocumentListenerContract;
+import com.cburch.logisim.gui.generic.EditorTheme;
 import com.cburch.logisim.gui.generic.OptionPane;
 import com.cburch.logisim.proj.Project;
 import com.cburch.logisim.util.FileUtil;
@@ -224,6 +225,7 @@ public class HdlContentEditor extends JDialog implements JInputDialog {
 
     RTextScrollPane sp = new RTextScrollPane(editor);
     sp.setFoldIndicatorEnabled(true);
+    EditorTheme.install(editor);
 
     add(sp, BorderLayout.CENTER);
     add(buttonsPanel, BorderLayout.SOUTH);